requirements.txt
python-telegram-bot~=20.7
httpx~=0.25.2
pymongo>=4.5,<5
python-dotenv~=1.0.1
openai~=1.12.0
pydantic~=2.6.1